requirements.txt
# ==========================================
# AI Video Dubber — Dependencies
# ==========================================
# Core pipeline requirements for Hindi dubbing
# Tested on Python 3.10+ with CUDA 11.8+
# ==========================================

# ------ Audio/Video Processing ------
ffmpeg-python>=0.2.0
librosa>=0.10.0
soundfile>=0.12.0
pydub>=0.25.1

# ------ Speech Recognition (Whisper) ------
openai-whisper>=20231117

# ------ Translation ------
# IndicTrans2 (primary — install via setup script)
# deep_translator is the lightweight fallback
deep-translator>=1.11.4

# ------ Voice Cloning (XTTS v2) ------
TTS>=0.22.0

# ------ Edge TTS (fallback, no GPU) ------
edge-tts>=6.1.9
nest-asyncio>=1.5.8

# ------ Face Enhancement (GFPGAN) ------
gfpgan>=1.3.8

# ------ Core ML ------
torch>=2.0.0
torchaudio>=2.0.0
torchvision>=0.15.0
transformers>=4.36.0
numpy>=1.24.0

# ------ Video Processing ------
opencv-python-headless>=4.8.0

# ------ Utilities ------
tqdm>=4.66.0
requests>=2.31.0
